requirements.txt
Flask==3.0.0
tensorflow-cpu==2.15.0
Pillow==10.0.0
numpy==1.26.0
gunicorn==21.2.0
